HSEARCH-3722 Rename hibernate-search-util-internal-integrationtest-elasticsearch to hibernate-search-util-internal-integrationtest-backend-elasticsearch
