+++ documentation/pom.xml
@@ -62,7 +62,7 @@
         </dependency>
         <dependency>
             <groupId>${project.groupId}</groupId>
-            <artifactId>hibernate-search-util-internal-integrationtest-elasticsearch</artifactId>
+            <artifactId>hibernate-search-util-internal-integrationtest-backend-elasticsearch</artifactId>
             <version>${project.version}</version>
             <scope>test</scope>
         </dependency>
