+++ integrationtest/pom.xml
@@ -36,7 +36,7 @@
             </dependency>
             <dependency>
                 <groupId>org.hibernate.search</groupId>
-                <artifactId>hibernate-search-util-internal-integrationtest-elasticsearch</artifactId>
+                <artifactId>hibernate-search-util-internal-integrationtest-backend-elasticsearch</artifactId>
                 <version>${project.version}</version>
             </dependency>
             <dependency>
