+++ util/internal/integrationtest/backend/elasticsearch/pom.xml
@@ -12,12 +12,12 @@
         <groupId>org.hibernate.search</groupId>
         <artifactId>hibernate-search-util-internal-integrationtest-parent</artifactId>
         <version>6.0.0-SNAPSHOT</version>
-        <relativePath>..</relativePath>
+        <relativePath>../..</relativePath>
     </parent>
-    <artifactId>hibernate-search-util-internal-integrationtest-elasticsearch</artifactId>
+    <artifactId>hibernate-search-util-internal-integrationtest-backend-elasticsearch</artifactId>
 
     <name>Hibernate Search Utils - Internal - Integration Tests - Elasticsearch</name>
-    <description>Hibernate Search integration testing utilities for tests involving Elasticsearch</description>
+    <description>Hibernate Search integration testing utilities for tests involving the Elasticsearch backend</description>
 
     <dependencies>
         <dependency>
