+++ util/internal/integrationtest/backend/elasticsearch/src/main/java/org/hibernate/search/util/impl/integrationtest/backend/elasticsearch/dialect/Elasticsearch60TestDialect.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.util.impl.integrationtest.elasticsearch.dialect;
+package org.hibernate.search.util.impl.integrationtest.backend.elasticsearch.dialect;
 
 import java.util.Arrays;
 import java.util.List;
@@ -13,10 +13,11 @@ import java.util.Optional;
 import org.hibernate.search.backend.elasticsearch.client.impl.Paths;
 import org.hibernate.search.backend.elasticsearch.util.spi.URLEncodedString;
 
+import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
 
 @SuppressWarnings("deprecation") // We use Paths.DOC on purpose
-public class Elasticsearch5TestDialect implements ElasticsearchTestDialect {
+public class Elasticsearch60TestDialect implements ElasticsearchTestDialect {
 
 	@Override
 	public boolean isEmptyMappingPossible() {
@@ -45,17 +46,9 @@ public class Elasticsearch5TestDialect implements ElasticsearchTestDialect {
 
 	@Override
 	public void setTemplatePattern(JsonObject object, String pattern) {
-		object.addProperty( "template", pattern );
-	}
-
-	@Override
-	public boolean isGeoPointIndexNullAsPossible() {
-		return false;
-	}
-
-	@Override
-	public boolean worksFineWithStrictGraterThanRangedQueriesOnScaledFloatField() {
-		return false;
+		JsonArray array = new JsonArray();
+		array.add( pattern );
+		object.add( "index_patterns", array );
 	}
 
 	@Override
